requirements.txt
Django==5.1.7
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.0
drf-yasg==1.21.7
python-dotenv==1.0.1
